Make riscof run

+++ core/src/zisk_inst_builder.rs
@@ -204,7 +204,7 @@ impl ZiskInstBuilder {
 
     /// Sets jump offsets.  The first offset is added to the pc when a set pc or a flag happens,
     /// and the second offset is the default one.
-    pub fn j(&mut self, j1: i32, j2: i32) {
+    pub fn j(&mut self, j1: i64, j2: i64) {
         self.i.jmp_offset1 = j1 as i64;
         self.i.jmp_offset2 = j2 as i64;
     }
